update my fork

+++ tests/test_r2s.py
@@ -390,7 +390,7 @@ def test_irradiation_setup_unstructured_nondef_tag():
         i += 1
 
 
-def _r2s_test_step1(r2s_run_dir):
+def _r2s_test_step1(r2s_run_dir, remove_step1_out=True):
     os.chdir(thisdir)
     # copy ../scripts/r2s.py to r2s_run_dir/r2s.py
     os.chdir("..")
@@ -419,10 +419,13 @@ def _r2s_test_step1(r2s_run_dir):
     f3 = filecmp.cmp(alara_fluxin, exp_alara_fluxin)
 
     # remove test output files
-    os.remove(alara_inp)
+    if remove_step1_out:
+        # these files are used in step2
+        os.remove(blank_mesh)
+        os.remove(alara_inp)
+    # these files are not used in step2
     os.remove(alara_fluxin)
     os.remove(alara_matlib)
-    os.remove(blank_mesh)
     os.remove(step1_file)
     os.remove(dst)
 
@@ -431,7 +434,7 @@ def _r2s_test_step1(r2s_run_dir):
     assert_equal(f3, True)
 
 
-def _r2s_test_step2(r2s_run_dir):
+def _r2s_test_step2(r2s_run_dir, remove_step1_out=True):
     os.chdir(thisdir)
     # copy ../scripts/r2s.py to r2s_run_dir/r2s.py
     os.chdir("..")
@@ -441,9 +444,10 @@ def _r2s_test_step2(r2s_run_dir):
 
     # output files of r2s step1
     alara_inp = os.path.join(r2s_run_dir, "alara_inp")
-    copyfile(os.path.join(r2s_run_dir, "exp_alara_inp"), alara_inp)
     blank_mesh = os.path.join(r2s_run_dir, "blank_mesh.h5m")
-    copyfile(os.path.join(r2s_run_dir, "exp_blank_mesh.h5m"), blank_mesh)
+    if remove_step1_out:
+        copyfile(os.path.join(r2s_run_dir, "exp_alara_inp"), alara_inp)
+        copyfile(os.path.join(r2s_run_dir, "exp_blank_mesh.h5m"), blank_mesh)
 
     # run r2s step2
     os.chdir(r2s_run_dir)
@@ -458,10 +462,26 @@ def _r2s_test_step2(r2s_run_dir):
     exp_e_bounds = os.path.join(r2s_run_dir, "exp_e_bounds")
     exp_t_p_src = os.path.join(
         r2s_run_dir, "exp_total_photon_source_intensities.txt")
+    exp_src_c1 = os.path.join(r2s_run_dir, "exp_source_1.h5m")
 
     # compare the results
     f4 = filecmp.cmp(e_bounds, exp_e_bounds)
     f5 = filecmp.cmp(t_p_src, exp_t_p_src)
+    f6 = True
+    # skip test if h5diff not exist
+    if 'unstructured' in r2s_run_dir:
+        ele_type = 'Tet4'
+    else:
+        ele_type = 'Hex8'
+    is_h5diff = os.system('which h5diff')
+    if is_h5diff == 0:
+        # compare two h5 files
+        command = ''.join(['h5diff --relative=1e-6 ', src_c1, ' ', exp_src_c1,
+            ' /tstt/elements/', ele_type, '/tags/source_density',
+            ' /tstt/elements/', ele_type, '/tags/source_density'])
+        diff_flag = os.system(command)
+        # return value 0 if no difference, 1 if differences found, 2 if error
+        f6 = True if diff_flag == 0 else False
 
     # remove test generated files
     os.remove(blank_mesh)
@@ -474,9 +494,10 @@ def _r2s_test_step2(r2s_run_dir):
 
     assert_equal(f4, True)
     assert_equal(f5, True)
+    assert_equal(f6, True)
 
 
-def test_r2s_script():
+def test_r2s_script_step_by_step():
 
     # skip test without dagmc
     try:
@@ -484,17 +505,43 @@ def test_r2s_script():
     except ImportError:
         raise SkipTest
 
+    remove_step1_out = True
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
     # test sub-voxel r2s
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "subvoxel_r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
     # test unstructured r2s
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "unstructured_r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
+
+
+def test_r2s_script():
+
+   # skip test without dagmc
+   try:
+       from pyne import dagmc
+   except ImportError:
+       raise SkipTest
+
+   remove_step1_out = False
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
+   # test sub-voxel r2s
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "subvoxel_r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
+   # test unstructured r2s
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "unstructured_r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
